init: shutdown the system on SIGRTMIN+4
This makes sysvinit behave nicely when it is running in a systemd-nspawn
container. When the user executes "machinectl stop", systemd sends
SIGRTMIN+4 to PID 1 in the container, and expects that to initiate a
graceful shutdown (power-off).

The cleanest way to replicate this in sysvinit is to execute the shutdown
command with appropriate options.